+++ pypit/arcoadd.py
@@ -521,12 +521,23 @@ def load_spec(files, iextensions=None, extract='opt', flux=True):
 
 
 def get_std_dev(irspec, ispec1d):
+    """
+    Parameters
+    ----------
+    irspec
+    ispec1d
+
+    Returns
+    -------
+
+    """
     # Only calculate on regions with 2 or more spectra
     msk = ~irspec.data['flux'].mask
     sum_msk = np.sum(msk, axis=0)
     gdp = sum_msk > 1
     # Here we go [note that dev_sig is still a masked array so we compress it after]
     dev_sig = (irspec.data['flux'][:,gdp] - ispec1d.flux[gdp]) / (irspec.data['sig'][:,gdp]**2 + ispec1d.sig[gdp]**2)
+    debugger.set_trace()
     std_dev = np.std(astropy.stats.sigma_clip(dev_sig.compressed(), sigma=5, iters=2))
     return std_dev, dev_sig.compressed()
 
@@ -571,8 +582,10 @@ def coadd_spectra(spectra, wave_grid_method='concatenate', niter=5,
     # Initial coadd
     spec1d = one_d_coadd(rspec, weights)
 
+    # Standard deviation
     std_dev, _ = get_std_dev(rspec, spec1d)
     msgs.info("Initial std_dev = {:g}".format(std_dev))
+    debugger.set_trace()
 
     iters = 0
     std_dev = 0.