+++ CHANGES.rst
@@ -6,7 +6,9 @@
 * Coadding; including docs; and tests
 * Now requiring astropy >= v1.3
 * raw_input handling for Python 3
-* coadd_ammendment
+* coadd handling of bad input
+* coadd bug fix
+
 0.7 (2017-02-07)
 ----------------